refactor(my-project): 优化用户信息展示和查询功能
-将 Account 类中的 id 类型从 Integer 改为 Long
- 在 AccountService 接口中添加 selectByIds 方法
- 在 AccountServiceImpl 类中实现 selectByIds 方法
- 优化 ProjectController 中的异常处理，统一返回系统异常信息
- 修改 ProjectFacade 中的项目查询接口，增加用户信息展示
- 更新前端 SkillMatch 组件，显示用户名称代替用户 ID

+++ my-project-backend/src/main/java/com/example/utils/JwtUtils.java
@@ -92,7 +92,7 @@ public class JwtUtils {
      * @param user 用户信息
      * @return 令牌
      */
-    public String createJwt(UserDetails user, String username, String nikeName, int userId, String role) {
+    public String createJwt(UserDetails user, String username, String nikeName, Long userId, String role) {
         if (this.frequencyCheck(userId)) {
             Algorithm algorithm = Algorithm.HMAC256(key);
             Date expire = this.expireTime();
@@ -167,7 +167,7 @@ public class JwtUtils {
      * @param userId 用户ID
      * @return 是否通过频率检测
      */
-    private boolean frequencyCheck(int userId) {
+    private boolean frequencyCheck(Long userId) {
         String key = Const.JWT_FREQUENCY + userId;
         return utils.limitOnceUpgradeCheck(key, limit_frequency, limit_base, limit_upgrade);
     }
